Split the tab-closing tests into their own file
useWorkspace.test.tsx reached 390 lines with them appended, over the 300-line
guard -- which CI caught and my local run did not, because I read the exit
status of an echo rather than the guard's.

The mock preamble is repeated rather than shared: `vi.mock` is hoisted per
file, so a shared harness module would have to re-declare it anyway.
