+++ apps/netscli-gui/src/workspace/useWorkspace.test.tsx
@@ -291,99 +291,3 @@ describe('tab lifecycle', () => {
     });
   });
 });
-
-// The tab context menu acts on the tab that was right-clicked, which is very
-// often not the active one. `closeOtherTabs` previously only ever kept the
-// active tab, so wiring a menu to it would have closed the tab the user was
-// pointing at whenever they right-clicked a background one.
-describe('closing a set of tabs', () => {
-  function fourTabs() {
-    const { result } = renderWorkspace();
-    act(() => {
-      result.current.closeAllTabs();
-    });
-    for (const kind of ['scan', 'ping', 'dns', 'trace'] as const) {
-      act(() => {
-        result.current.addTab(kind);
-      });
-    }
-    return result;
-  }
-
-  it('keeps the tab it was given, not the active one', () => {
-    const result = fourTabs();
-    const [first, second] = result.current.tabs;
-    act(() => {
-      result.current.setActiveTabId(result.current.tabs[3].id);
-    });
-
-    act(() => {
-      result.current.closeOtherTabs(second.id);
-    });
-
-    expect(result.current.tabs.map((tab) => tab.id)).toEqual([second.id]);
-    expect(result.current.activeTabId).toBe(second.id);
-    expect(result.current.tabs.map((tab) => tab.id)).not.toContain(first.id);
-  });
-
-  it('closes only what is to the right, and keeps the anchor', () => {
-    const result = fourTabs();
-    const ids = result.current.tabs.map((tab) => tab.id);
-
-    act(() => {
-      result.current.closeTabsBeside(ids[1], 'right');
-    });
-
-    expect(result.current.tabs.map((tab) => tab.id)).toEqual([ids[0], ids[1]]);
-  });
-
-  it('closes only what is to the left, and keeps the anchor', () => {
-    const result = fourTabs();
-    const ids = result.current.tabs.map((tab) => tab.id);
-
-    act(() => {
-      result.current.closeTabsBeside(ids[2], 'left');
-    });
-
-    expect(result.current.tabs.map((tab) => tab.id)).toEqual([ids[2], ids[3]]);
-  });
-
-  it('moves the selection to the anchor when the active tab was closed', () => {
-    const result = fourTabs();
-    const ids = result.current.tabs.map((tab) => tab.id);
-    act(() => {
-      result.current.setActiveTabId(ids[3]);
-    });
-
-    act(() => {
-      result.current.closeTabsBeside(ids[1], 'right');
-    });
-
-    expect(result.current.activeTabId).toBe(ids[1]);
-  });
-
-  it('leaves the active tab alone when it survived', () => {
-    const result = fourTabs();
-    const ids = result.current.tabs.map((tab) => tab.id);
-    act(() => {
-      result.current.setActiveTabId(ids[0]);
-    });
-
-    act(() => {
-      result.current.closeTabsBeside(ids[1], 'right');
-    });
-
-    expect(result.current.activeTabId).toBe(ids[0]);
-  });
-
-  it('does nothing when there is nothing on that side', () => {
-    const result = fourTabs();
-    const ids = result.current.tabs.map((tab) => tab.id);
-
-    act(() => {
-      result.current.closeTabsBeside(ids[0], 'left');
-    });
-
-    expect(result.current.tabs.map((tab) => tab.id)).toEqual(ids);
-  });
-});
